modif pour post

+++ src/main/java/fr/epsi/b3devc1/msprapi/repository/CountryRepository.java
@@ -3,5 +3,5 @@ package fr.epsi.b3devc1.msprapi.repository;
 import fr.epsi.b3devc1.msprapi.model.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {  // Changement ici (Integer → Long)
 }
